detailsPanel, navigPanel, buttonPanel and searchPanels all renamed to createDetailPanel() etc... - more appropriate method name.
Constants interface created to hold all the app constants.

Duplicate code in createDialogPanel() in AddRecordDialog and createDetailsPanel() in ApplicationPresenter. Three methods extracted from createDialogPanel, the first to build empDetails and the second to create the buttonPanel and finally the third to add the fonts and listeners. The first two were common to two classes and these were placed in ApplicationPresenter, the third was left in AddRecordDialog as this was unique to this class,